Modifying libcoroutine so that it accepts our Valgrind settings and doesn't make warnings

+++ src/coroutine/Coro.cc
@@ -48,7 +48,7 @@
 	Coro *c = (coro); \
 		c->valgrindStackId = VALGRIND_STACK_REGISTER( \
 											 c->stack, \
-											 c->stack + c->requestedStackSize); \
+											 (size_t)c->stack + c->requestedStackSize); \
 }
 
 #define STACK_DEREGISTER(coro) \
